move plain car blockstore to separate module
